pyproject.toml
[tool.poetry]
name = "weather-api"
version = "0.1.0"
description = ""
authors = ["Alexey Kudimov <work.alexeykudimov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
asyncio = "^3.4.3"
aiohttp = "^3.8.1"
pytest = "^7.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
